+++ Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("NiumaClaw.Team")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.0.1.0")]
-[assembly: AssemblyInformationalVersion("1.0.1+online.20260608115731")]
+[assembly: AssemblyFileVersion("1.0.2.0")]
+[assembly: AssemblyInformationalVersion("1.0.2+mac-client.20260701")]
 [assembly: AssemblyProduct("NiumaClaw.Team")]
 [assembly: AssemblyTitle("NiumaClaw.Team")]
-[assembly: AssemblyVersion("1.0.1.0")]
+[assembly: AssemblyVersion("1.0.2.0")]